SPR1-2674 Silence remaining NumPy type warnings
NumPy >= 1.20 complains if you use the deprecated np.int, np.float and
np.object aliases. Replace those with straight int, float and object
types.

Pytest reports: 211 passed, 10 skipped, 8 warnings in 60.05s

+++ katdal/test/test_chunkstore_s3.py
@@ -338,7 +338,7 @@ class TestS3ChunkStore(ChunkStoreTestBase):
 
     def test_missing_or_empty_buckets(self):
         slices = (slice(0, 1),)
-        dtype = np.dtype(np.float)
+        dtype = np.dtype(float)
         # Without create_array the bucket is missing
         with assert_raises(StoreUnavailable):
             self.store.get_chunk(f'{BUCKET}-missing/x', slices, dtype)
